groupe1/requirements.txt
Flask==1.1.4
Flask-MySQLdb==0.2.0
Flask-JWT-Extended==4.2.0
bcrypt==3.2.0
mysql-connector-python==8.0.28
torch==2.0.1
torchvision==0.15.0
markupsafe==1.1.1